pyproject.toml
[project]
name = "job-scraper"
version = "0.1.0"
description = "Job scraper that collects job postings which fit my filters and sends notifications to Discord when it finds them."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "importlib>=1.0.4",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
    "ty>=0.0.24",
]
